Deprecate v1 clients
