+++ client/kafka/kafka_test.go
@@ -1,102 +1,96 @@
 package kafka
 
 import (
+	"context"
+	"fmt"
+	"strings"
 	"testing"
 
+	"github.com/Shopify/sarama"
+	"github.com/beatlabs/patron/correlation"
+	"github.com/beatlabs/patron/trace"
+	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/ext"
+	"github.com/opentracing/opentracing-go/mocktracer"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func TestNewMessage(t *testing.T) {
-	m := NewMessage("TOPIC", []byte("TEST"))
-	assert.Equal(t, "TOPIC", m.topic)
-	assert.Equal(t, []byte("TEST"), m.body)
-}
-
-func TestNewMessageWithKey(t *testing.T) {
-	tests := []struct {
-		name    string
-		data    []byte
-		key     string
-		wantErr bool
-	}{
-		{name: "success", data: []byte("TEST"), key: "TEST"},
-		{name: "failure due to empty message key", data: []byte("TEST"), key: "", wantErr: true},
+func TestBuilder_Create(t *testing.T) {
+	t.Parallel()
+	type args struct {
+		brokers []string
+		cfg     *sarama.Config
 	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := NewMessageWithKey("TOPIC", tt.data, tt.key)
-			if tt.wantErr {
-				assert.Error(t, err)
-				assert.Nil(t, got)
-			} else {
-				assert.NoError(t, err)
-				assert.NotNil(t, got)
-			}
+	tests := map[string]struct {
+		args        args
+		expectedErr string
+	}{
+		"missing brokers": {args: args{brokers: nil, cfg: sarama.NewConfig()}, expectedErr: "brokers are empty or have an empty value\n"},
+		"missing config":  {args: args{brokers: []string{"123"}, cfg: nil}, expectedErr: "no Sarama configuration specified\n"},
+	}
+	for name, tt := range tests {
+		tt := tt
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+			got, err := New(tt.args.brokers, tt.args.cfg).Create()
+
+			require.EqualError(t, err, tt.expectedErr)
+			require.Nil(t, got)
 		})
 	}
 }
 
-func TestNewAsyncProducer_Failure(t *testing.T) {
-	got, chErr, err := NewBuilder([]string{}).CreateAsync()
-	assert.Error(t, err)
-	assert.Nil(t, got)
-	assert.Nil(t, chErr)
-}
-
-func TestNewAsyncProducer_Option_Failure(t *testing.T) {
-	got, chErr, err := NewBuilder([]string{"xxx"}).WithVersion("xxxx").CreateAsync()
-	assert.Error(t, err)
-	assert.Nil(t, got)
-	assert.Nil(t, chErr)
-}
-
-func TestNewSyncProducer_Failure(t *testing.T) {
-	got, err := NewBuilder([]string{}).CreateSync()
-	assert.Error(t, err)
-	assert.Nil(t, got)
-}
-
-func TestNewSyncProducer_Option_Failure(t *testing.T) {
-	got, err := NewBuilder([]string{"xxx"}).WithVersion("xxxx").CreateSync()
-	assert.Error(t, err)
-	assert.Nil(t, got)
-}
-
-func TestNewMessageWithHeader(t *testing.T) {
-	tests := []struct {
-		name                 string
-		data                 []byte
-		setHeaderKeys        []string
-		setHeaderValues      []string
-		expectedHeaderKeys   []string
-		expectedHeaderValues []string
-	}{
-		{
-			name: "2-headers", data: []byte("TEST"),
-			setHeaderKeys: []string{"header1", "header2"}, setHeaderValues: []string{"value1", "value2"},
-			expectedHeaderKeys: []string{"header1", "header2"}, expectedHeaderValues: []string{"value1", "value2"},
-		},
-		{
-			name: "2-headers", data: []byte("TEST"),
-			setHeaderKeys: []string{"header1", "header1"}, setHeaderValues: []string{"value1", "value2"},
-			expectedHeaderKeys: []string{"header1", "header1"}, expectedHeaderValues: []string{"value1", "value2"},
-		},
+func TestBuilder_CreateAsync(t *testing.T) {
+	t.Parallel()
+	type args struct {
+		brokers []string
+		cfg     *sarama.Config
 	}
+	tests := map[string]struct {
+		args        args
+		expectedErr string
+	}{
+		"missing brokers": {args: args{brokers: nil, cfg: sarama.NewConfig()}, expectedErr: "brokers are empty or have an empty value\n"},
+		"missing config":  {args: args{brokers: []string{"123"}, cfg: nil}, expectedErr: "no Sarama configuration specified\n"},
+	}
+	for name, tt := range tests {
+		tt := tt
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+			got, chErr, err := New(tt.args.brokers, tt.args.cfg).CreateAsync()
 
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			msg := NewMessage("TOPIC", tt.data)
-
-			// set message headers
-			for i := 0; i < len(tt.setHeaderKeys); i++ {
-				msg.SetHeader(tt.setHeaderKeys[i], tt.setHeaderValues[i])
-			}
-
-			// verify
-			for i := 0; i < len(tt.expectedHeaderKeys); i++ {
-				assert.Equal(t, string(msg.headers[i].Key), tt.expectedHeaderKeys[i])
-				assert.Equal(t, string(msg.headers[i].Value), tt.expectedHeaderValues[i])
-			}
+			require.EqualError(t, err, tt.expectedErr)
+			require.Nil(t, got)
+			require.Nil(t, chErr)
 		})
 	}
+}
+
+func TestDefaultProducerSaramaConfig(t *testing.T) {
+	sc, err := DefaultProducerSaramaConfig("name", true)
+	require.NoError(t, err)
+	require.True(t, strings.HasSuffix(sc.ClientID, fmt.Sprintf("-%s", "name")))
+	require.True(t, sc.Producer.Idempotent)
+
+	sc, err = DefaultProducerSaramaConfig("name", false)
+	require.NoError(t, err)
+	require.False(t, sc.Producer.Idempotent)
+}
+
+func Test_injectTracingAndCorrelationHeaders(t *testing.T) {
+	mtr := mocktracer.New()
+	opentracing.SetGlobalTracer(mtr)
+	t.Cleanup(func() { mtr.Reset() })
+	ctx := correlation.ContextWithID(context.Background(), "123")
+	sp, _ := trace.ChildSpan(context.Background(), trace.ComponentOpName(componentTypeAsync, "topic"), componentTypeAsync,
+		ext.SpanKindProducer, asyncTag, opentracing.Tag{Key: "topic", Value: "topic"})
+	msg := sarama.ProducerMessage{}
+	assert.NoError(t, injectTracingAndCorrelationHeaders(ctx, &msg, sp))
+	assert.Len(t, msg.Headers, 4)
+	assert.Equal(t, correlation.HeaderID, string(msg.Headers[0].Key))
+	assert.Equal(t, "123", string(msg.Headers[0].Value))
+	assert.Equal(t, "mockpfx-ids-traceid", string(msg.Headers[1].Key))
+	assert.Equal(t, "mockpfx-ids-spanid", string(msg.Headers[2].Key))
+	assert.Equal(t, "mockpfx-ids-sampled", string(msg.Headers[3].Key))
 }
